# Add block cache